Move remaining generator functions into the generators directory
