+++ packages/core/lib/policeGateway/leds/LedsGateway.ts
@@ -21,10 +21,10 @@ import { asnQueryResponseSchema } from "../../../schemas/leds/asnQueryResponse"
 import Asn from "../../Asn"
 import PoliceApiError from "../PoliceApiError"
 import endpoints from "./endpoints"
-import generateCheckName from "./generateCheckName"
-import generateRequestHeaders from "./generateRequestHeaders"
+import generateCheckName from "./generators/generateCheckName"
 import { generateDisposalResultsRequest } from "./generators/generateDisposalResultsRequest"
 import { generateRemandRequest } from "./generators/generateRemandRequest"
+import generateRequestHeaders from "./generators/generateRequestHeaders"
 import { generateSubsequentDisposalRequest } from "./generators/generateSubsequentDisposalRequest"
 import mapToPoliceQueryResult from "./mappers/mapToPoliceQueryResult"
