+++ packages/core/lib/policeGateway/leds/generators/generateCheckName.test.ts
