+++ packages/core/lib/policeGateway/leds/generators/generateCheckName.ts
